wip(vscode) Update vscode extension to use synchronous messaging and fs
