+++ utopia-vscode-common/src/index.ts
@@ -1,11 +1,7 @@
 export * from './path-utils'
-export * from './mailbox'
-export * from './messages'
-export * from './fs/fs-types'
-export * from './fs/fs-utils'
 export * from './prettier-utils'
+export * from './messages-to-utopia'
+export * from './messages-to-vscode'
 export * from './utopia-vscode-config'
-export * from './vscode-communication'
-export * from './window-messages'
 
 export const ProjectIDPlaceholderPrefix = 'PLACEHOLDER_DURING_LOADING'
