+++ utopia-vscode-extension/src/utopia-fs.ts
@@ -24,31 +24,39 @@ import {
   Position,
   Range,
   workspace,
+  commands,
 } from 'vscode'
 import {
-  watch,
-  stopWatching,
   stat,
   pathIsDirectory,
   createDirectory,
   readFile,
   exists,
   writeFile,
-  appendToPath,
-  dirname,
-  stripRootPrefix,
   deletePath,
   rename,
   getDescendentPaths,
   isDirectory,
   readDirectory,
-  RootDir,
   readFileSavedContent,
   writeFileSavedContent,
   readFileSavedContentAsUTF8,
   pathIsFileWithUnsavedContent,
+  readFileAsUTF8,
+  writeFileAsUTF8,
+  getItem,
+  createDirectoryWithoutError,
+  isFile,
+  isNotDirectoryError,
+} from './in-mem-fs'
+import {
+  appendToPath,
+  dirname,
+  RootDir,
+  type ProjectFile,
+  vsCodeFileDelete,
 } from 'utopia-vscode-common'
-import { fromUtopiaURI, toUtopiaURI } from './path-utils'
+import { addSchemeToPath, allPathsUpToPath } from './path-utils'
 
 interface EventQueue<T> {
   queue: T[]
@@ -70,8 +78,6 @@ export class UtopiaFSExtension
   private disposable: Disposable
 
   private fileChangeEventQueue = newEventQueue<FileChangeEvent>()
-  private utopiaSavedChangeEventQueue = newEventQueue<Uri>()
-  private utopiaUnsavedChangeEventQueue = newEventQueue<Uri>()
 
   private allFilePaths: string[] | null = null
 
@@ -89,13 +95,9 @@ export class UtopiaFSExtension
 
   // FileSystemProvider
   readonly onDidChangeFile: Event<FileChangeEvent[]> = this.fileChangeEventQueue.emitter.event
-  readonly onUtopiaDidChangeSavedContent: Event<Uri[]> =
-    this.utopiaSavedChangeEventQueue.emitter.event
-  readonly onUtopiaDidChangeUnsavedContent: Event<Uri[]> =
-    this.utopiaUnsavedChangeEventQueue.emitter.event
 
-  private queueEvent<T>(event: T, eventQueue: EventQueue<T>): void {
-    eventQueue.queue.push(event)
+  private queueEvents<T>(events: Array<T>, eventQueue: EventQueue<T>): void {
+    eventQueue.queue.push(...events)
 
     if (eventQueue.handle != null) {
       clearTimeout(eventQueue.handle)
@@ -107,82 +109,96 @@ export class UtopiaFSExtension
     }, 5)
   }
 
-  private queueFileChangeEvent(event: FileChangeEvent): void {
+  private queueFileChangeEvents(events: Array<FileChangeEvent>): void {
     this.clearCachedFiles()
-    this.queueEvent(event, this.fileChangeEventQueue)
+    this.queueEvents(events, this.fileChangeEventQueue)
   }
 
-  private queueUtopiaSavedChangeEvent(resource: Uri): void {
-    this.queueEvent(resource, this.utopiaSavedChangeEventQueue)
-  }
-
-  private queueUtopiaUnsavedChangeEvent(resource: Uri): void {
-    this.queueEvent(resource, this.utopiaUnsavedChangeEventQueue)
-  }
-
-  private async notifyFileChanged(path: string, modifiedBySelf: boolean): Promise<void> {
-    const uri = toUtopiaURI(this.projectID, path)
-    const hasUnsavedContent = await pathIsFileWithUnsavedContent(path)
+  private notifyFileChanged(path: string) {
+    const uri = addSchemeToPath(this.projectID, path)
+    const hasUnsavedContent = pathIsFileWithUnsavedContent(path)
     const fileWasSaved = !hasUnsavedContent
 
     if (fileWasSaved) {
       // Notify VS Code of updates to the saved content
-      this.queueFileChangeEvent({
-        type: FileChangeType.Changed,
-        uri: uri,
-      })
-    }
-
-    if (!modifiedBySelf) {
-      // Notify our extension of changes coming from Utopia only
-      if (fileWasSaved) {
-        this.queueUtopiaSavedChangeEvent(uri)
-      } else {
-        this.queueUtopiaUnsavedChangeEvent(uri)
-      }
+      this.queueFileChangeEvents([
+        {
+          type: FileChangeType.Changed,
+          uri: uri,
+        },
+      ])
     }
   }
 
-  private notifyFileCreated(path: string): void {
-    this.queueFileChangeEvent({
-      type: FileChangeType.Created,
-      uri: toUtopiaURI(this.projectID, path),
-    })
-  }
-
-  private notifyFileDeleted(path: string): void {
-    this.queueFileChangeEvent({
-      type: FileChangeType.Deleted,
-      uri: toUtopiaURI(this.projectID, path),
-    })
-  }
-
-  watch(uri: Uri, options: { recursive: boolean; excludes: string[] }): Disposable {
-    const path = fromUtopiaURI(uri)
-    watch(
-      path,
-      options.recursive,
-      this.notifyFileCreated.bind(this),
-      this.notifyFileChanged.bind(this),
-      this.notifyFileDeleted.bind(this),
-      () => {
-        /* no op */
+  private notifyFileCreated(path: string) {
+    const parentDirectory = dirname(path)
+    this.queueFileChangeEvents([
+      {
+        type: FileChangeType.Created,
+        uri: addSchemeToPath(this.projectID, path),
       },
-    )
-
-    return new Disposable(() => {
-      stopWatching(path, options.recursive)
-    })
+      {
+        type: FileChangeType.Changed,
+        uri: addSchemeToPath(this.projectID, parentDirectory),
+      },
+    ])
   }
 
-  async exists(uri: Uri): Promise<boolean> {
-    const path = fromUtopiaURI(uri)
+  private notifyFileDeleted(path: string) {
+    const parentDirectory = dirname(path)
+    this.queueFileChangeEvents([
+      {
+        type: FileChangeType.Deleted,
+        uri: addSchemeToPath(this.projectID, path),
+      },
+      {
+        type: FileChangeType.Changed,
+        uri: addSchemeToPath(this.projectID, parentDirectory),
+      },
+    ])
+  }
+
+  private notifyFileRenamed(oldPath: string, newPath: string) {
+    const oldParentDirectory = dirname(oldPath)
+    const newParentDirectory = dirname(newPath)
+    const parentChanged = oldParentDirectory !== newParentDirectory
+    this.queueFileChangeEvents([
+      {
+        type: FileChangeType.Deleted,
+        uri: addSchemeToPath(this.projectID, oldPath),
+      },
+      {
+        type: FileChangeType.Created,
+        uri: addSchemeToPath(this.projectID, newPath),
+      },
+      {
+        type: FileChangeType.Changed,
+        uri: addSchemeToPath(this.projectID, oldParentDirectory),
+      },
+      ...(parentChanged
+        ? [
+            {
+              type: FileChangeType.Changed,
+              uri: addSchemeToPath(this.projectID, newParentDirectory),
+            },
+          ]
+        : []),
+    ])
+  }
+
+  watch(): Disposable {
+    // No need for this since all events are manually fired
+    return new Disposable(() => {})
+  }
+
+  exists(uri: Uri): boolean {
+    const path = uri.path
     return exists(path)
   }
 
-  async stat(uri: Uri): Promise<FileStat> {
-    const path = fromUtopiaURI(uri)
-    const stats = await stat(path)
+  stat(uri: Uri): FileStat {
+    const path = uri.path
+    const stats = stat(path)
     const fileType = isDirectory(stats) ? FileType.Directory : FileType.File
 
     return {
@@ -193,36 +209,30 @@ export class UtopiaFSExtension
     }
   }
 
-  async readDirectory(uri: Uri): Promise<[string, FileType][]> {
-    const path = fromUtopiaURI(uri)
-    const children = await readDirectory(path)
-    const result: Promise<[string, FileType]>[] = children.map((childName) =>
-      pathIsDirectory(appendToPath(path, childName)).then((resultIsDirectory) => [
-        childName,
-        resultIsDirectory ? FileType.Directory : FileType.File,
-      ]),
-    )
-    return Promise.all(result)
+  readDirectory(uri: Uri): Array<[string, FileType]> {
+    const path = uri.path
+    const children = readDirectory(path)
+    return children.map((childName) => {
+      const resultIsDirectory = pathIsDirectory(appendToPath(path, childName))
+      return [childName, resultIsDirectory ? FileType.Directory : FileType.File]
+    })
   }
 
-  async createDirectory(uri: Uri): Promise<void> {
-    const path = fromUtopiaURI(uri)
-    await createDirectory(path)
+  createDirectory(uri: Uri) {
+    const path = uri.path
+    createDirectory(path)
+    this.notifyFileCreated(path)
   }
 
-  async readFile(uri: Uri): Promise<Uint8Array> {
-    const path = fromUtopiaURI(uri)
+  readFile(uri: Uri): Uint8Array {
+    const path = uri.path
     return readFileSavedContent(path)
   }
 
-  async writeFile(
-    uri: Uri,
-    content: Uint8Array,
-    options: { create: boolean; overwrite: boolean },
-  ): Promise<void> {
-    const path = fromUtopiaURI(uri)
+  writeFile(uri: Uri, content: Uint8Array, options: { create: boolean; overwrite: boolean }) {
+    const path = uri.path
+    const fileExists = exists(path)
     if (!options.create || !options.overwrite) {
-      const fileExists = await exists(path)
       if (!fileExists && !options.create) {
         throw FileSystemError.FileNotFound(uri)
       } else if (fileExists && !options.overwrite) {
@@ -230,74 +240,135 @@ export class UtopiaFSExtension
       }
     }
 
-    await writeFileSavedContent(path, content)
+    writeFileSavedContent(path, content)
+    if (fileExists) {
+      this.notifyFileChanged(path)
+    } else {
+      this.notifyFileCreated(path)
+    }
   }
 
-  async delete(uri: Uri, options: { recursive: boolean }): Promise<void> {
-    const path = fromUtopiaURI(uri)
-    await deletePath(path, options.recursive)
+  ensureDirectoryExists(pathToEnsure: string) {
+    const allPaths = allPathsUpToPath(pathToEnsure)
+    let createdDirectories: Array<string> = []
+    for (const pathToCreate of allPaths) {
+      const existingNode = getItem(pathToCreate)
+      if (existingNode == null) {
+        createDirectoryWithoutError(pathToCreate)
+        createdDirectories.push(pathToCreate)
+      } else if (isFile(existingNode)) {
+        throw isNotDirectoryError(pathToCreate)
+      }
+    }
+
+    createdDirectories.forEach((createdDirectory) => this.notifyFileCreated(createdDirectory))
   }
 
-  async rename(oldUri: Uri, newUri: Uri, options: { overwrite: boolean }): Promise<void> {
-    const oldPath = fromUtopiaURI(oldUri)
-    const newPath = fromUtopiaURI(newUri)
+  writeProjectFile(projectFile: ProjectFile) {
+    switch (projectFile.type) {
+      case 'PROJECT_DIRECTORY': {
+        const { filePath } = projectFile
+        this.ensureDirectoryExists(filePath)
+        break
+      }
+      case 'PROJECT_TEXT_FILE': {
+        const { filePath, savedContent, unsavedContent } = projectFile
+        const fileExists = exists(filePath)
+        const alreadyExistingFile = fileExists ? readFileAsUTF8(filePath) : null
+        const fileDiffers =
+          alreadyExistingFile == null ||
+          alreadyExistingFile.content !== savedContent ||
+          alreadyExistingFile.unsavedContent !== unsavedContent
+        if (fileDiffers) {
+          // Avoid pushing a file to the file system if the content hasn't changed.
+          writeFileAsUTF8(filePath, savedContent, unsavedContent)
+
+          if (fileExists) {
+            this.notifyFileChanged(filePath)
+          } else {
+            this.notifyFileCreated(filePath)
+          }
+        }
+        break
+      }
+      default:
+        const _exhaustiveCheck: never = projectFile
+        throw new Error(`Invalid file projectFile type ${projectFile}`)
+    }
+  }
+
+  delete(uri: Uri, options: { recursive: boolean }) {
+    this.silentDelete(uri.path, options)
+    commands.executeCommand('utopia.toUtopiaMessage', vsCodeFileDelete(uri.path))
+  }
+
+  silentDelete(path: string, options: { recursive: boolean }) {
+    deletePath(path, options.recursive)
+    this.notifyFileDeleted(path)
+  }
+
+  rename(oldUri: Uri, newUri: Uri, options: { overwrite: boolean }) {
+    const oldPath = oldUri.path
+    const newPath = newUri.path
 
     if (!options.overwrite) {
-      const fileExists = await exists(newPath)
+      const fileExists = exists(newPath)
       if (fileExists) {
         throw FileSystemError.FileExists(newUri)
       }
     }
 
-    await rename(oldPath, newPath)
+    rename(oldPath, newPath)
+    this.notifyFileRenamed(oldPath, newPath)
   }
 
-  async copy(source: Uri, destination: Uri, options: { overwrite: boolean }): Promise<void> {
+  copy(source: Uri, destination: Uri, options: { overwrite: boolean }) {
     // It's not clear where this will ever be called from, but it seems to be from the side bar
     // that isn't available in Utopia, so this implementation is "just in case"
-    const sourcePath = fromUtopiaURI(source)
-    const destinationPath = fromUtopiaURI(destination)
+    const sourcePath = source.path
+    const destinationPath = destination.path
     const destinationParentDir = dirname(destinationPath)
-    const destinationParentDirExists = await exists(destinationParentDir)
+    const destinationParentDirExists = exists(destinationParentDir)
 
     if (!destinationParentDirExists) {
-      throw FileSystemError.FileNotFound(toUtopiaURI(this.projectID, destinationParentDir))
+      throw FileSystemError.FileNotFound(addSchemeToPath(this.projectID, destinationParentDir))
     }
 
     if (!options.overwrite) {
-      const destinationExists = await exists(destinationPath)
+      const destinationExists = exists(destinationPath)
       if (destinationExists && !options.overwrite) {
         throw FileSystemError.FileExists(destination)
       }
     }
 
-    const { content, unsavedContent } = await readFile(sourcePath)
-    await writeFile(destinationPath, content, unsavedContent)
+    const { content, unsavedContent } = readFile(sourcePath)
+    writeFile(destinationPath, content, unsavedContent)
+    this.notifyFileCreated(destinationPath)
   }
 
   // FileSearchProvider
 
-  async provideFileSearchResults(
+  provideFileSearchResults(
     query: FileSearchQuery,
     options: FileSearchOptions,
     _token: CancellationToken,
-  ): Promise<Uri[]> {
+  ): Array<Uri> {
     // TODO Support all search options
-    const { result: foundPaths } = await this.filterFilePaths(query.pattern, options.maxResults)
-    return foundPaths.map((p) => toUtopiaURI(this.projectID, p))
+    const { result: foundPaths } = this.filterFilePaths(query.pattern, options.maxResults)
+    return foundPaths.map((p) => addSchemeToPath(this.projectID, p))
   }
 
   // TextSearchProvider
 
-  async provideTextSearchResults(
+  provideTextSearchResults(
     query: TextSearchQuery,
     options: TextSearchOptions,
     progress: Progress<TextSearchResult>,
     token: CancellationToken,
-  ): Promise<TextSearchComplete> {
+  ): TextSearchComplete {
     // This appears to only be callable from the side bar that isn't available in Utopia
     // TODO Support all search options
-    const { result: filePaths, limitHit } = await this.filterFilePaths(options.includes[0])
+    const { result: filePaths, limitHit } = this.filterFilePaths(options.includes[0])
 
     if (filePaths.length > 0) {
       for (const filePath of filePaths) {
@@ -305,7 +376,7 @@ export class UtopiaFSExtension
           break
         }
 
-        const content = await readFileSavedContentAsUTF8(filePath)
+        const content = readFileSavedContentAsUTF8(filePath)
 
         const lines = splitIntoLines(content)
         for (let i = 0; i < lines.length; i++) {
@@ -313,7 +384,7 @@ export class UtopiaFSExtension
           const index = line.indexOf(query.pattern)
           if (index !== -1) {
             progress.report({
-              uri: toUtopiaURI(this.projectID, filePath),
+              uri: addSchemeToPath(this.projectID, filePath),
               ranges: new Range(
                 new Position(i, index),
                 new Position(i, index + query.pattern.length),
@@ -336,11 +407,11 @@ export class UtopiaFSExtension
 
   // Common
 
-  private async filterFilePaths(
+  private filterFilePaths(
     query: string | undefined,
     maxResults?: number,
-  ): Promise<{ result: string[]; limitHit: boolean }> {
-    const filePaths = await this.getAllPaths()
+  ): { result: Array<string>; limitHit: boolean } {
+    const filePaths = this.getAllPaths()
     let result: string[] = []
     let limitHit = false
     let remainingCount = maxResults == null ? Infinity : maxResults
@@ -352,7 +423,7 @@ export class UtopiaFSExtension
         break
       }
 
-      if (!pattern || pattern.exec(stripRootPrefix(path))) {
+      if (!pattern || pattern.exec(path)) {
         if (remainingCount === 0) {
           // We've already found the max number of results, but we want to flag that there are more
           limitHit = true
@@ -369,9 +440,9 @@ export class UtopiaFSExtension
     }
   }
 
-  async getAllPaths(): Promise<string[]> {
+  getAllPaths(): Array<string> {
     if (this.allFilePaths == null) {
-      const result = await getDescendentPaths(RootDir)
+      const result = getDescendentPaths(RootDir)
       this.allFilePaths = result
       return result
     } else {
